+++ src/app/post/new/_components/postForm/postFormElement/index.ts
@@ -1,3 +1,9 @@
+export * from './postFormLabel';
+
 export * from './postImageField';
 
 export * from './postPlaceField';
+
+export * from './postKeywordField';
+
+export * from './postMenuField';
