+++ src/app/post/new/_components/postForm/postFormElement/postKeywordField/index.ts
@@ -1,0 +1,1 @@
+export * from './PostKeywordField';
