+++ src/app/post/new/_components/postForm/postFormElement/postMenuField/MenuTag.tsx
@@ -1,0 +1,20 @@
+import { ReactNode } from 'react';
+import { CloseSVG } from '@/components/svg/CloseSVG';
+
+export const MenuTag = ({
+  children,
+  onClick,
+}: {
+  children: ReactNode;
+  onClick: () => void;
+}) => {
+  return (
+    <span
+      className='inline-flex h-[45px] cursor-pointer items-center justify-center gap-1 rounded-3xl bg-gray-50 px-5 text-gray-800 body-16 hover:bg-gray-75 [&>svg]:h-4 [&>svg]:w-4'
+      onClick={onClick}
+    >
+      <span>{children}</span>
+      <CloseSVG />
+    </span>
+  );
+};
